requirements.txt
# Microsoft Agent Framework - pinned to known working version
agent-framework==1.0.0b260210

# Azure Monitoring - Use Distro for automatic instrumentation
azure-monitor-opentelemetry

# Azure Identity
azure-identity

# Environment
python-dotenv

# Weather API
requests

# OpenAI SDK (for LLM code generation in chart tool)
openai

# Charting (for local chart generation)
matplotlib
